refactor: rename utils/kpf_utils.py to utils/kpf.py
Pure module rename plus its test file (test_kpf_utils.py -> test_kpf.py);
all imports updated. No behavior change.
