+++ kpfpipe/data_models/masters/base.py
@@ -17,7 +17,7 @@ import pandas as pd
 
 from kpfpipe.data_models.base import KPFDataModel
 from kpfpipe.utils.io import kpf_filename
-from kpfpipe.utils.kpf_utils import get_obs_id
+from kpfpipe.utils.kpf import get_obs_id
 
 logger = logging.getLogger(__name__)
