+++ kpfpipe/quality_control/diagnostics/level1.py
@@ -8,7 +8,7 @@ master relative to the observation time.
 from datetime import datetime
 
 from kpfpipe.quality_control.diagnostics.base import Diagnostics
-from kpfpipe.utils.kpf_utils import get_timestamp, kpf_timestamp_to_datetime
+from kpfpipe.utils.kpf import get_timestamp, kpf_timestamp_to_datetime
 
 # Master-calibration age metrics: the RECEIPT path keyword (written by
 # CalibrationAssociation) -> the age keyword whose signed (master - obs) value
